feat(irs-registry-client):[#202] Removed beans that required spring profile from irs-registry-client
